Testing for routers
Working on passing multiple routes at once
