+++ public/javascripts/searchClientside.js
@@ -1,5 +1,4 @@
-function setAction(form) {
-    form.action = "register.html";
-    alert(form.action);
-    return false;
-    }
+
+function helloWorld() {
+    console.log("Hello World");
+}
