+++ routes/index.js
@@ -2,8 +2,11 @@ var express = require('express');
 var router = express.Router();
 
 /* GET home page. */
-router.get('/', function(req, res, next) {
+router.get('/', async function(req, res, next) {
+  var test = await req.baseUrl;
+  console.log('id: ' + test);
   res.render('index', { layout: 'index' });
 });
+
 
 module.exports = router;
